+++ R/crop_rain_area_from_radolan.R
@@ -1,6 +1,8 @@
 # crop_area_from_radolan_stack -------------------------------------------------
-crop_area_from_radolan_stack <- function(area, radolan_stack)
+crop_area_from_radolan_stack <- function(area, radolan_stack, use_mask = TRUE)
 {
+  #kwb.utils::assignPackageObjects("fhpredict")
+
   # Convert the area list structure to a matrix with columns "lon" and "lat"
   lonlat <- area_to_longitude_latitude_matrix(area)
 
@@ -12,8 +14,8 @@ crop_area_from_radolan_stack <- function(area, radolan_stack)
     crs_to = as.character(raster::crs(radolan_stack))
   )
 
-  # Crop the polygon areas from the raster stack
-  raster::crop(x = radolan_stack, y = polygon)
+  # Crop or mask the polygon areas from the raster stack
+  crop_or_mask(radolan_stack, polygon, use_mask)
 }
 
 # area_to_longitude_latitude_matrix --------------------------------------------